pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "schools-scraper"
version = "0.1.0"
description = "Web scraper for schools data with analysis and GPT integration"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "duckdb>=0.9.0",
    "pandas>=2.0.0",
    "openai>=1.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "schedule>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
schools-scraper = "schools_scraper.cli:app"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

